Renamed and relocated Google Analytics component
